Rename package url to s3url
