+++ core/op_s3.go
@@ -15,8 +15,8 @@ import (
 
 	"github.com/peak/s5cmd/op"
 	"github.com/peak/s5cmd/opt"
+	"github.com/peak/s5cmd/s3url"
 	"github.com/peak/s5cmd/stats"
-	"github.com/peak/s5cmd/url"
 )
 
 func S3Copy(job *Job, wp *WorkerParams) (stats.StatType, error) {
@@ -155,7 +155,7 @@ func S3BatchDownload(job *Job, wp *WorkerParams) (stats.StatType, error) {
 
 		arg1 := NewJobArgument(
 			"s3://"+job.args[0].s3.Bucket+"/"+*li.Key,
-			&url.S3Url{Bucket: job.args[0].s3.Bucket, Key: *li.Key},
+			&s3url.S3Url{Bucket: job.args[0].s3.Bucket, Key: *li.Key},
 		)
 
 		var dstFn string
@@ -346,7 +346,7 @@ func S3BatchCopy(job *Job, wp *WorkerParams) (stats.StatType, error) {
 
 		arg1 := NewJobArgument(
 			"s3://"+job.args[0].s3.Bucket+"/"+*li.Key,
-			&url.S3Url{Bucket: job.args[0].s3.Bucket, Key: *li.Key},
+			&s3url.S3Url{Bucket: job.args[0].s3.Bucket, Key: *li.Key},
 		)
 
 		var dstFn string
@@ -358,7 +358,7 @@ func S3BatchCopy(job *Job, wp *WorkerParams) (stats.StatType, error) {
 
 		arg2 := NewJobArgument(
 			"s3://"+job.args[1].s3.Bucket+"/"+job.args[1].s3.Key+dstFn,
-			&url.S3Url{Bucket: job.args[1].s3.Bucket, Key: job.args[1].s3.Key + dstFn},
+			&s3url.S3Url{Bucket: job.args[1].s3.Bucket, Key: job.args[1].s3.Key + dstFn},
 		)
 
 		subJob := job.MakeSubJob(subCmd, op.Copy, []*JobArgument{arg1, arg2}, job.opts)
